+++ examples/counter.py
@@ -34,8 +34,6 @@ app = nu.With(
     ),
 )
 
-print(app)
-
 
 if __name__ == "__main__":
     import asyncio